+++ db/updates/update_11.9-12.0.sql
@@ -5,3 +5,4 @@ ALTER TABLE `inventory_accessories` CHANGE `supplier_link` `supplier_link` VARCH
 ALTER TABLE `inventory_accessories` CHANGE `supplier` `supplier` VARCHAR(255) NOT NULL; 
 ALTER TABLE `inventory_accessories` ADD UNIQUE(`name`);
 ALTER TABLE `inventory_compounds` ADD UNIQUE(`name`);
+ALTER TABLE `bottles` ADD UNIQUE(`name`);