+++ js/import.bottles.js
@@ -13,7 +13,7 @@ function uploadProgressHandler(event) {
 function loadHandler(event) {
 	$("#status").html(event.target.responseText);
 	$(".progress").hide();
-	//$("#btnImportAccessories").hide();
+	//$("#btnImportBottles").hide();
 	$("#jsonFile").val('');
 	$('#btnCloseBK').prop('value', 'Close');
 	$("#uploadProgressBar").css("width", "0%");
@@ -28,7 +28,7 @@ function abortHandler(event) {
 }
 	
 $(".progress").hide();
-$("#btnImportAccessories").prop("disabled", true);
+$("#btnImportBottles").prop("disabled", true);
 $("#jsonFile").change(function(){
 	var allowedTypes = ['application/json'];
 	var file = this.files[0];
@@ -39,23 +39,23 @@ $("#jsonFile").change(function(){
 	if(!allowedTypes.includes(fileType)){
 		$("#JSRestMsg").html('<div class="alert alert-info">Invalid file selected. Please select a JSON file exported from PV.</div>');
 		$("#jsonFile").val('');
-		$("#btnImportAccessories").prop("disabled", true);
+		$("#btnImportBottles").prop("disabled", true);
 		return false;
 	}
 	
 	if (fileSize > fileSizePHP){
 		$("#JSRestMsg").html('<div class="alert alert-info">File size <strong>('+formatBytes(fileSize)+')</strong> is exceeding your server file upload limit '+ formatBytes(fileSizePHP)+'</div>');
 		$("#jsonFile").val('');
-		$("#btnImportAccessories").prop("disabled", true);
+		$("#btnImportBottles").prop("disabled", true);
 		return false;
 	}
 	
-	$("#btnImportAccessories").prop("disabled", false);
-	$('#btnImportAccessories').prop('value', 'Import');
+	$("#btnImportBottles").prop("disabled", false);
+	$('#btnImportBottles').prop('value', 'Import');
 });
 
 
-$('#btnImportAccessories').click(function() {
+$('#btnImportBottles').click(function() {
 	
 	event.preventDefault();
 	var fd = new FormData();
@@ -66,7 +66,7 @@ $('#btnImportAccessories').click(function() {
 	}
 	
 	$.ajax({ 
-		url: '/core/core.php?action=importAccessories', 
+		url: '/core/core.php?action=importBottles', 
 		type: 'POST',
 		data: fd,
 		contentType: false,
@@ -80,32 +80,32 @@ $('#btnImportAccessories').click(function() {
             xhr.addEventListener("error", errorHandler, false);
             xhr.addEventListener("abort", abortHandler, false);
 			$(".progress").show();
-			$("#btnImportAccessories").prop("disabled", true);
-			$('#btnImportAccessories').prop('value', 'Please wait...');
+			$("#btnImportBottles").prop("disabled", true);
+			$('#btnImportBottles').prop('value', 'Please wait...');
         	return xhr;
         },	
 		success: function (data) {
 			if(data.success){
 				var msg = '<div class="alert alert-success"><i class="fa-solid fa-circle-check mx-2"></i>'+data.success+'</div>';
-				$("#btnImportAccessories").hide();
+				$("#btnImportBottles").hide();
 				$("#backupArea").css('display', 'none');
-				$('#tdDataAccessories').DataTable().ajax.reload(null, true);
+				$('#tdDataBottles').DataTable().ajax.reload(null, true);
 
 			}else if(data.error){
 				var msg = '<div class="alert alert-danger"><i class="fa-solid fa-triangle-exclamation mx-2"></i>'+data.error+'</div>';
-				$("#btnImportAccessories").show();
-				$("#btnImportAccessories").prop("disabled", false);
-				$('#btnImportAccessories').prop('value', 'Import');
+				$("#btnImportBottles").show();
+				$("#btnImportBottles").prop("disabled", false);
+				$('#btnImportBottles').prop('value', 'Import');
 			}
-			$('#btnImportAccessories').prop('value', 'Import');
-			$("#btnImportAccessories").prop("disabled", false);
+			$('#btnImportBottles').prop('value', 'Import');
+			$("#btnImportBottles").prop("disabled", false);
 			$('#JSRestMsg').html(msg);
 		},
 		error: function (xhr, status, error) {
 			$('#JSRestMsg').html('<div class="alert alert-danger"><i class="fa-solid fa-triangle-exclamation mx-2"></i>An ' + status + ' occurred, check server logs for more info. ' + error + '</div>');
-			$("#btnImportAccessories").show();
-			$("#btnImportAccessories").prop("disabled", false);
-			$('#btnImportAccessories').prop('value', 'Import');
+			$("#btnImportBottles").show();
+			$("#btnImportBottles").prop("disabled", false);
+			$('#btnImportBottles').prop('value', 'Import');
 		}
 		
 	  });