+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Migrating backend scripts under a common backend api
 - Import json functions update
 - Added import for accessories
+- Added import for bottles
 
 ### Version 11.9
 - Added system logs access via the UI for docker/cloud installations - this comes disabled by default